fix enabling bluetooth zrfam swap

+++ my-hype.sh
@@ -94,25 +94,46 @@ fi
 
 git config --global core.editor "nano"
 
-# Enable Bluetooth
-enable_bluetooth() {
-	echo "Enabling Bluetooth..."
-	sudo cp /etc/bluetooth/main.conf /etc/bluetooth/main.conf.backup
-	if sudo sed -i 's/#AutoEnable=true/AutoEnable=true/' /etc/bluetooth/main.conf; then
-		sudo systemctl start bluetooth.service
-		sudo systemctl enable bluetooth.service
-		echo "Bluetooth configuration successful."
-	else
-		echo "Failed to enable Bluetooth. Exiting script."
-		exit 1
-	fi
+# Prompt user to answer yes or no
+ask_user() {
+	while true; do
+		read -rp "$1 (y/n): " yn
+		case $yn in
+			[Yy]* ) return 0;;
+			[Nn]* ) return 1;;
+			* ) echo "Please answer y or n.";;
+		esac
+	done
 }
 
+# Enable Bluetooth
+clear
+if ask_user "Do you want to enable Bluetooth?"; then
+	if [ -f /etc/bluetooth/main.conf ]; then
+		echo "Enabling Bluetooth..."
+		sudo cp /etc/bluetooth/main.conf /etc/bluetooth/main.conf.backup
+		if sudo sed -i 's/#AutoEnable=true/AutoEnable=true/' /etc/bluetooth/main.conf; then
+			sudo systemctl start bluetooth.service
+			sudo systemctl enable bluetooth.service
+			echo "Bluetooth configuration successful."
+		else
+			echo "Failed to enable Bluetooth. Exiting script."
+			exit 1
+		fi
+	else
+		echo "Bluetooth configuration file not found. Exiting script."
+		exit 1
+	fi
+else
+	echo "Skipping Bluetooth configuration."
+fi
+
 # Configure zram swap
-configure_zram() {
-	echo "Enabling zram..."
-	sudo cp /etc/systemd/zram-generator.conf /etc/systemd/zram-generator.conf.backup
-	if sudo tee /etc/systemd/zram-generator.conf >/dev/null <<EOF
+if ask_user "Do you want to configure zram swap?"; then
+	if [ -f /etc/systemd/zram-generator.conf ]; then
+		echo "Enabling zram..."
+		sudo cp /etc/systemd/zram-generator.conf /etc/systemd/zram-generator.conf.backup
+		if sudo tee /etc/systemd/zram-generator.conf >/dev/null <<EOF
 [zram0]
 zram-size = ram
 EOF
@@ -121,19 +142,14 @@ then
 else
 	echo "Failed to configure zram. Exiting script."
 	exit 1
+		fi
+	else
+		echo "zram configuration file not found. Exiting script."
+		exit 1
 	fi
-}
-
-read -rp "Do you want to enable Bluetooth? (yes/no): " enable_bluetooth_prompt
-if [ "$enable_bluetooth_prompt" = "yes" ]; then
-	enable_bluetooth
+else
+	echo "Skipping zram configuration."
 fi
-
-read -rp "Do you want to configure zram=ram zram? (yes/no): " configure_zram_prompt
-if [ "$configure_zram_prompt" = "yes" ]; then
-	configure_zram
-fi
-
 
 install_sddm_theme() {
 	if [ -f /usr/lib/sddm/sddm.conf.d/default.conf ]; then
@@ -154,6 +170,7 @@ install_sddm_theme() {
 }
 
 # Desktop Manager Setup (optional)
+clear
 echo "Would you like to install the SDDM login manager?"
 echo "Note: If you're using GNOME or KDE desktop, skip this step."
 echo "Only install SDDM if you are installing on vanilla arch"
